moving task specific stimuli to designated module
