+++ rewardgym/psychopy_render/hcp.py
@@ -1,91 +1,6 @@
-import importlib.util
-
-import numpy as np
-
-if importlib.util.find_spec("psychopy") is not None:
-    from psychopy.visual import TextStim
-    from psychopy.visual.rect import Rect
-else:
-    from .psychopy_stubs import TextStim, Rect
-
 from .default_images import fixation_cross
-from .stimuli import ActionStimulus, BaseStimulus, FeedBackStimulus, ImageStimulus
-
-
-class ShowCard(BaseStimulus):
-    def __init__(
-        self,
-        text,
-        condition_text,
-        duration=1.0,
-        width=250,
-        height=350,
-        lineWidth=3,
-        lineColor="white",
-        fillColor="grey",
-        textColor="white",
-        target="reward",
-        font_height=150,
-        position=(0, 0),
-        name=None,
-    ):
-        super().__init__(name=name, duration=duration)
-
-        self.height = height
-        self.width = width
-        self.lineWidth = lineWidth
-        self.lineColor = lineColor
-        self.fillColor = fillColor
-        self.textColor = textColor
-        self.text = text
-        self.target = target
-        self.condition_text = condition_text
-        self.font_height = font_height
-
-    def setup(self, win, **kwargs):
-        self.textStim = TextStim(
-            win=win,
-            name=self.name + "_text",
-            text=self.text,
-            color=self.textColor,
-            height=self.font_height,
-        )
-
-        self.rectStim = Rect(
-            win=win,
-            width=self.width,
-            height=self.height,
-            fillColor=self.fillColor,
-            lineWidth=self.lineWidth,
-            lineColor=self.lineColor,
-            name=self.name + "_rect",
-        )
-
-    def display(self, win, logger, reward, action, **kwargs):
-        if self.target == "action":
-            reward = action
-
-        logger.key_strokes(win)
-
-        stim_onset = logger.get_time()
-
-        card = np.random.choice(self.condition_text[reward])
-        display_text = self.text.format(card)
-        self.textStim.setText(display_text)
-
-        self.rectStim.autoDraw = True
-        self.textStim.autoDraw = True
-        win.flip()
-
-        self.rectStim.autoDraw = False
-        self.textStim.autoDraw = False
-
-        logger.wait(win, self.duration, stim_onset)
-
-        logger.log_event(
-            {"event_type": self.name, "expected_duration": self.duration},
-            onset=stim_onset,
-        )
+from .special_stimuli import TextWithBorder
+from .stimuli import ActionStimulus, FeedBackStimulus, ImageStimulus
 
 
 def get_info_dict(seed=None, key_dict={"left": 0, "right": 1}, **kwargs):
@@ -108,7 +23,7 @@ def get_info_dict(seed=None, key_dict={"left": 0, "right": 1}, **kwargs):
     info_dict = {
         0: {
             "psychopy": [
-                ShowCard(
+                TextWithBorder(
                     "{0}",
                     {0: ["?"]},
                     name="response-cue",
@@ -120,7 +35,7 @@ def get_info_dict(seed=None, key_dict={"left": 0, "right": 1}, **kwargs):
         1: {
             "psychopy": [
                 wait_after_response,
-                ShowCard(
+                TextWithBorder(
                     "{0}",
                     condition_text={1: [1, 2, 3, 4], -0.5: [6, 7, 8, 9], 0: [5]},
                     name="outcome",
@@ -132,7 +47,7 @@ def get_info_dict(seed=None, key_dict={"left": 0, "right": 1}, **kwargs):
         2: {
             "psychopy": [
                 wait_after_response,
-                ShowCard(
+                TextWithBorder(
                     "{0}",
                     condition_text={-0.5: [1, 2, 3, 4], 1: [6, 7, 8, 9], 0: [5]},
                     name="outcome",
